requirements.txt
Flask==2.1.2
Flask-restful==0.3.9
flask_cors=3.0.10
pymongo==3.12.0
requests-toolbelt==0.9.1
numpy==1.22.3
onnx==1.11.0
onnxruntime==1.11.1
pillow==9.0.1
instafilter==0.3.1
